requirements.txt
flask==2.3.2
openai==0.27.8
boto3==1.26.0
gunicorn==20.1.0
flask-cors==4.0.0
azure-cognitiveservices-speech
google-genai
base64